backend/requirements.txt
fastapi
uvicorn
motor
pymongo
python-dotenv
pydantic
pydantic-settings
PyPDF2
Pillow
python-multipart
passlib[bcrypt]
httpx
pyjwt
